refactor routers to be implemented in typescript
