+++ src/routes/info.ts
@@ -1,0 +1,10 @@
+import express from "express";
+const router = express.Router();
+
+import { handleQueryAttributes } from "../middleware/attributes";
+import { handleQueryEventNames } from "../middleware/events";
+
+router.get("/eventNames", handleQueryEventNames);
+router.get("/attributes", handleQueryAttributes);
+
+export default router;
